Added script to generate ons-oa-geo-london geojson files
Note: This uses the new V2 boundaries for the 2011 Output Areas, published July 24, 2020, while the analysis was conducted using the previous version, which is no longer available. A brief analysis comparing the file created using this code and the file used in the analysis suggests that the differences are minimal. The geometric difference of the areas is almost always below 0.01%, with 14 of 25053 geometries with a difference above 0.1% and one geometry (E00022402) with the area of the difference of geometries equal to 12% of the new geometry (V2).